code review

+++ core/src/main/java/io/questdb/griffin/engine/groupby/MonthTimestampMicrosSampler.java
@@ -59,7 +59,7 @@ public class MonthTimestampMicrosSampler implements TimestampSampler {
     @Override
     public long nextTimestamp(long timestamp, long numSteps) {
         try {
-            return addMonth(timestamp, (int) Math.multiplyExact(numSteps, stepMonths));
+            return addMonth(timestamp, Math.toIntExact(Math.multiplyExact(numSteps, stepMonths)));
         } catch (ArithmeticException e) {
             return Long.MAX_VALUE;
         }
