+++ core/src/main/java/io/questdb/griffin/engine/groupby/YearTimestampMicrosSampler.java
@@ -58,7 +58,7 @@ public class YearTimestampMicrosSampler implements TimestampSampler {
     @Override
     public long nextTimestamp(long timestamp, long numSteps) {
         try {
-            return addYears(timestamp, (int) Math.multiplyExact(numSteps, stepYears));
+            return addYears(timestamp, Math.toIntExact(Math.multiplyExact(numSteps, stepYears)));
         } catch (ArithmeticException e) {
             return Long.MAX_VALUE;
         }
